api/requirements.txt
fastapi
uvicorn
sentence-transformers
torch
pydantic
psycopg2-binary
python-dotenv
